fix(frontend): improve error handling and Suspense wrapping
- Extract getErrorMessage utility with proper type guards (CR-003)
- Wrap SignupPage with Suspense boundary like LoginPage (CR-006)
- Add documentation comment to parseJwtPayload (CR-001)
