+++ frontend/src/app/signup/page.tsx
@@ -1,128 +1,10 @@
-"use client";
-
-import { useState } from "react";
-import Link from "next/link";
-import { Input, Button } from "@/components/ui";
-import { GuestGuard } from "@/components/auth";
-import { useSignup } from "@/hooks";
-import type { ApiError } from "@/types";
-import { AxiosError } from "axios";
+import { Suspense } from "react";
+import { SignupForm } from "./signup-form";
 
 export default function SignupPage() {
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [errors, setErrors] = useState<Record<string, string>>({});
-
-  const signup = useSignup();
-
-  function validate(): boolean {
-    const newErrors: Record<string, string> = {};
-    if (!name) newErrors.name = "Name is required";
-    else if (name.length > 50)
-      newErrors.name = "Name must be 50 characters or less";
-    if (!email) newErrors.email = "Email is required";
-    else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email))
-      newErrors.email = "Invalid email format";
-    if (!password) newErrors.password = "Password is required";
-    else if (password.length < 8)
-      newErrors.password = "Password must be at least 8 characters";
-    if (password !== confirmPassword)
-      newErrors.confirmPassword = "Passwords do not match";
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  }
-
-  function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    if (!validate()) return;
-    signup.mutate({ email, password, name });
-  }
-
-  const serverError =
-    signup.error instanceof AxiosError
-      ? (signup.error.response?.data as { error?: ApiError })?.error?.message ??
-        "Signup failed. Please try again."
-      : signup.error
-        ? "An unexpected error occurred."
-        : null;
-
   return (
-    <GuestGuard>
-    <div className="flex min-h-[70vh] items-center justify-center px-4">
-      <div className="w-full max-w-sm">
-        <h1 className="text-center text-2xl font-bold text-gray-900">
-          Create your account
-        </h1>
-        <p className="mt-2 text-center text-sm text-gray-600">
-          Already have an account?{" "}
-          <Link
-            href="/login"
-            className="font-medium text-primary-600 hover:text-primary-500"
-          >
-            Sign in
-          </Link>
-        </p>
-
-        {serverError && (
-          <div className="mt-4 rounded-lg bg-red-50 p-3 text-sm text-red-700">
-            {serverError}
-          </div>
-        )}
-
-        <form onSubmit={handleSubmit} className="mt-8 space-y-5">
-          <Input
-            id="name"
-            label="Name"
-            type="text"
-            placeholder="Your name"
-            autoComplete="name"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            error={errors.name}
-          />
-          <Input
-            id="email"
-            label="Email"
-            type="email"
-            placeholder="you@example.com"
-            autoComplete="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            error={errors.email}
-          />
-          <Input
-            id="password"
-            label="Password"
-            type="password"
-            placeholder="At least 8 characters"
-            autoComplete="new-password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            error={errors.password}
-          />
-          <Input
-            id="confirmPassword"
-            label="Confirm Password"
-            type="password"
-            placeholder="Re-enter your password"
-            autoComplete="new-password"
-            value={confirmPassword}
-            onChange={(e) => setConfirmPassword(e.target.value)}
-            error={errors.confirmPassword}
-          />
-
-          <Button
-            type="submit"
-            isLoading={signup.isPending}
-            className="w-full"
-          >
-            Create Account
-          </Button>
-        </form>
-      </div>
-    </div>
-    </GuestGuard>
+    <Suspense>
+      <SignupForm />
+    </Suspense>
   );
 }
