+++ frontend/src/lib/index.ts
@@ -1,2 +1,3 @@
 export { apiClient } from "./api-client";
 export { getQueryClient, makeQueryClient } from "./query-client";
+export { getErrorMessage } from "./error";
